requirements.txt
pygame==2.5.2
Pillow==10.2.0
mutagen==1.47.0
pyinstaller==6.3.0 